Rename requireFns to getComponentFunctions
