+++ src/utils/components/getComponentsFromStateFile.js
@@ -1,6 +1,7 @@
 const { keys, reduce, not, isEmpty, append, pickBy, prop } = require('ramda')
 const reduceIndexed = require('../reduceIndexed')
 const getComponentFunctions = require('./getComponentFunctions')
+const getComponentRootPath = require('./getComponentRootPath')
 const getChildrenIds = require('./getChildrenIds')
 const getState = require('../state/getState')
 const getInputs = require('../state/getInputs')
@@ -17,6 +18,7 @@ function getComponentsFromStateFile(stateFile) {
   return reduce((accum, id) => {
     const component = stateFile[id]
     const { type } = component
+    const componentRoot = getComponentRootPath(type)
     return {
       ...accum,
       [id]: {
@@ -28,7 +30,7 @@ function getComponentsFromStateFile(stateFile) {
         dependencies: [], // TODO: do we need to determine the dependencies here?
         children: getChildrenIds(component),
         promise: deferredPromise(),
-        fns: getComponentFunctions(type)
+        fns: getComponentFunctions(componentRoot)
       }
     }
   }, {}, componentIds)
